Implement random usernames for anonymous players
Replace socket id with either the registered users username
or randomized animal names for anonymous users

Adjust backend socket io logic to allow use of usernames

Adjust frontend on-connect event to read username
