--- flask/app/socketio_bp/events.py
@@ -1,167 +1,0 @@
-import os
-import random
-import uuid
-from pathlib import Path
-from collections import deque
-
-from flask import request, session
-from flask_socketio import send, emit, rooms, join_room, leave_room
-
-from app.socketio_init import socketio
-
-# keep track of connected clients for reference in queue
-clients = set()
-
-# use queue to push clients into rooms
-queue = deque()
-
-# use animal names and number as name for random
-animals_count = {}
-
-with open(os.path.join(Path(__file__).resolve().parent, "animals.txt")) as f:
-    animals = f.read().splitlines()
-
-
-def get_anon_name(choices, choices_count):
-    # returns a randomly chosen name from a given list
-    # increments the name in a tracker and appends the count number to the name
-    # e.g., dog#1234
-    choice = random.choice(choices)
-    choices_count[choice] = choices_count.get(choice, 0) + 1
-    return f"{choice}#{choices_count[choice]}"
-
-
-@socketio.on("connect")
-def handle_connect(auth):
-    sid = request.sid
-
-    # TODO confirm that this works for logged in users
-    username = session.get("username", sid)
-
-    emit("general notification", f"{username} CONNECTED", broadcast=True)
-
-    clients.add(sid)
-
-
-@socketio.on("disconnect")
-def handle_disconnect():
-    sid = request.sid
-    clients.remove(sid)
-
-    for room in rooms():
-        emit("room notification", f"{sid} DISCONNECTED", to=room)
-
-
-@socketio.on("leave")
-def on_leave(data):
-
-    sid = request.sid
-    room = data["room"]
-
-    leave_room(room)
-
-    # notify user of successful exit
-    emit("user notification", f"Left {room}")
-    emit("status change", {"status": "CONNECTED"})
-
-    # notify room that a user has left
-    emit("room notification", f"{sid} LEFT", to=room)
-
-
-# queue/join room
-@socketio.on("queue")
-def handle_queue():
-    sid = request.sid
-
-    # prevent duplicate queuing
-    if len(rooms()) > 1:
-        return emit("user notification", "Already in a room!")
-
-    queue.append(sid)
-    players = set()
-
-    while len(players) < 2:
-        try:
-            popped = queue.popleft()
-
-            # handle disconnected clients
-            if popped in clients:
-                players.add(popped)
-
-        except IndexError:
-            break
-
-    # failed to find match for last player, requeue
-    if len(players) < 2:
-        last = players.pop()
-        queue.append(last)
-        return emit(
-            "user notification", "Waiting for players to join...", to=last
-        )
-
-    # match found
-    player1 = players.pop()
-    player2 = players.pop()
-
-    room_id = uuid.uuid4().hex
-
-    # TODO investigate if this is necessary
-    # check for collision with existing socket ids
-    while room_id in clients:
-        room_id = uuid.uuid4().hex
-
-    join_room(room_id, sid=player1)
-    join_room(room_id, sid=player2)
-
-    emit("room notification", f"{player1} JOINED", to=room_id)
-    emit("room notification", f"{player2} JOINED", to=room_id)
-
-    emit("user notification", f"Joined room {room_id}", to=player1)
-    emit("user notification", f"Opponent: {player2}", to=player1)
-    emit("joined room", {"room": room_id, "opponent": player2}, to=player1)
-
-    emit("user notification", f"Joined room {room_id}", to=player2)
-    emit("user notification", f"Opponent: {player1}", to=player2)
-    emit("joined room", {"room": room_id, "opponent": player1}, to=player2)
-
-
-# TODO: implement frontend
-# cancel queue
-@socketio.on("cancel queue")
-def handle_cancel_queue(data):
-
-    sid = request.sid
-
-    try:
-        queue.remove(sid)
-        emit("cancel queue")
-    except ValueError:
-        emit("user notification", "Not in queue")
-
-
-@socketio.on("choice")
-def handle_choice(data):
-    try:
-        room = rooms()[1]
-        emit("choice", data, to=room, include_self=False)
-    except IndexError:
-        emit("ERROR", "Server could not determine if you are in a room")
-
-
-@socketio.on("general chat")
-def handle_general_chat(data):
-    sid = request.sid
-    username = session.get("username", sid)
-    data["username"] = username
-    send(data, broadcast=True)
-
-
-@socketio.on("room chat")
-def handle_room_chat(data):
-    # send named event so frontend can handle separately from general messages
-    # check if client is in room
-    if data["room"] in rooms():
-        sid = request.sid
-        username = session.get("username", sid)
-        data["username"] = username
-        emit("room chat", data, to=data["room"])
