+++ flask/app/socketio_bp/utils/animals.txt
